refactor(utils): share one escapeHtml instead of three copies
The five-character HTML escaper existed verbatim in three places:
`grounding-render.ts` (exported), `mcp-oauth-callback.ts` (private), and
`test/mcp/mcp-oauth-callback.test.ts`, whose comment said outright that it
re-implemented the source copy "and verify parity" because the real one was
unreachable.

Move it to `src/utils/html-entities.ts` — a leaf module both call sites can
import, and the natural home next to the existing `decodeHtmlEntities`. Both
tests now exercise the shared function.

No behaviour change: the moved body is byte-identical to both copies.
